+++ src/bika/lims/controlpanel/bika_analysisservices.py
@@ -170,49 +170,39 @@ class AnalysisServicesView(BikaListingView):
                 "index": "sortable_title",
                 "replace_url": "getURL",
                 "sortable": self.can_sort}),
-            ("Unit", {
-                "title": _("Unit"),
-                "sortable": False}),
             ("Keyword", {
                 "title": _("Keyword"),
                 "index": "getKeyword",
                 "attr": "getKeyword",
                 "sortable": self.can_sort}),
-            ("SortKey", {
-                "title": _("Sort Key"),
-                "sortable": False}),
+            ("Category", {
+                "title": _("Category"),
+                "attr": "getCategoryTitle",
+                "sortable": self.can_sort}),
             ("Methods", {
                 "title": _("Methods"),
                 "sortable": self.can_sort}),
-            ("Instruments", {
-                "title": _("Instruments"),
+            ("Department", {
+                "title": _("Department"),
+                "toggle": False,
+                "sortable": self.can_sort}),
+            ("Unit", {
+                "title": _("Unit"),
+                "sortable": False}),
+            ("Price", {
+                "title": _("Price"),
+                "sortable": self.can_sort}),
+            ("MaxTimeAllowed", {
+                "title": _("Max Time"),
+                "toggle": False,
+                "sortable": self.can_sort}),
+            ("DuplicateVariation", {
+                "title": _("Dup Var"),
                 "toggle": False,
                 "sortable": False}),
             ("Calculation", {
                 "title": _("Calculation"),
                 "sortable": False}),
-            ("MaxTimeAllowed", {
-                "title": _("Maximum turnaround time"),
-                "toggle": False,
-                "sortable": self.can_sort}),
-            ("DecimalPrecision", {
-                "title": _("Decimal Precision"),
-                "toggle": False,
-                "sortable": False}),
-            ("StringResult", {
-                "title": _("String Result"),
-                "sortable": self.can_sort}),
-            ("LDL", {
-                "title": _("LDL"),
-                "toggle": False,
-                "sortable": False}),
-            ("UDL", {
-                "title": _("UDL"),
-                "toggle": False,
-                "sortable": False}),
-            ("DuplicateVariation", {
-                "title": _("Duplicate Variation"),
-                "toggle": False,}),
             ("SortKey", {
                 "ajax": True,
                 "help": _(
@@ -221,25 +211,6 @@ class AnalysisServicesView(BikaListingView):
                 "refetch": True,
                 "title": _("Sort Key"),
                 "sortable": False}),
-            ("Accredited", {
-                "title": _("Accredited"),
-                "toggle": False,
-                "sortable": False}),
-            ("Category", {
-                "title": _("Category"),
-                "attr": "getCategoryTitle",
-                "sortable": self.can_sort}),
-            ("Department", {
-                "title": _("Department"),
-                "toggle": False,
-                "sortable": self.can_sort}),
-            ("Price", {
-                "title": _("Price"),
-                "toggle": False,
-                "sortable": self.can_sort}),
-            ("BulkPrice", {
-                "title": _("Bulk Price"),
-                "sortable": self.can_sort}),
         ))
 
         copy_transition = {
@@ -273,14 +244,6 @@ class AnalysisServicesView(BikaListingView):
         if not self.context.bika_setup.getShowPrices():
             for i in range(len(self.review_states)):
                 self.review_states[i]["columns"].remove("Price")
-                self.review_states[i]["columns"].remove("BulkPrice")
-            del self.columns['Price']
-            del self.columns['BulkPrice']
-
-    def can_edit(self, service):
-        """Check if manage is allowed
-        """
-        return check_permission(ModifyPortalContent, self.context)
 
     def can_edit(self, service):
         """Check if manage is allowed
@@ -379,13 +342,7 @@ class AnalysisServicesView(BikaListingView):
             item["MaxTimeAllowed"] = self.format_maxtime(maxtime)
 
         # Price
-        user_roles = api.get_current_user().getRoles()
-        if "Manager" in user_roles:
-            financial_permissions = True
-        else:
-            financial_permissions = False
-        if financial_permissions:
-            item["Price"] = self.format_price(obj.Price)
+        item["Price"] = self.format_price(obj.Price)
 
         # Duplicate Variation
         dup_variation = obj.DuplicateVariation
@@ -408,50 +365,6 @@ class AnalysisServicesView(BikaListingView):
         sortkey = obj.getSortKey()
         item["SortKey"] = sortkey
 
-        #LDL and UDL
-        ldl = obj.getLowerDetectionLimit()
-        udl = obj.getUpperDetectionLimit()
-        if not ldl and not udl:
-            pass
-        elif(udl=='0' and ldl=='0'):
-            pass
-        else:
-            if isinstance(udl,tuple) and isinstance(ldl,tuple):
-                item["LDL"] = ldl[0] if ldl else 0
-                item["UDL"] = udl[0] if udl else 0
-            else:
-                item["LDL"] = ldl if ldl else 0
-                item["UDL"] = udl if udl else 0
-
-        #Decimal Precision
-        decimal_precision = obj.Precision
-        item["DecimalPrecision"] = decimal_precision
-
-        #Instruments
-        instruments = obj.getInstruments()
-        instrument_names = []
-        for instrument in instruments:
-            instrument_names.append(instrument.title)
-        item["Instruments"] = instrument_names
-    
-        #Accredited
-        accredited = obj.getAccredited()
-        if accredited:
-            item["Accredited"] = "Y"
-        else:
-            item["Accredited"] = ""
-
-        #Bulk Price
-        bulk_price = obj.BulkPrice
-        if financial_permissions:
-            item["BulkPrice"] = self.format_price(bulk_price)
-
-        #String result Y/N
-        string_result = obj.getStringResult()
-        if string_result:
-            item["StringResult"] = "Y"
-        else:
-            item["StringResult"] = ""
         if self.can_edit(obj):
             item["allow_edit"].append("SortKey")